requirements.txt
numpy==1.20.2
pandas==1.2.4
python-dateutil==2.8.1
pytz==2021.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.15.0
tqdm==4.60.0
